requirements.txt
python-dotenv
discord
dragonmapper
